Fixes for some serious bugs, addition of much more detailed tests.
It turns out there was a mistake in the application of a window which caused serious issues with most inner products.
This is now fixed.

There are now much more comprehensive tests on some of the likelihood and inner product code, both for noisy and noise-free examples
using a trivial waveform.